extract common stuff, jitpack
