--- src/main/kotlin/works/szabope/plugins/common/CommonBundle.kt
@@ -1,21 +1,0 @@
-package works.szabope.plugins.common
-
-import com.intellij.DynamicBundle
-import org.jetbrains.annotations.NonNls
-import org.jetbrains.annotations.PropertyKey
-
-@NonNls
-private const val BUNDLE = "messages.works_szabope_plugins_CommonBundle"
-
-object CommonBundle {
-
-    private val bundle = DynamicBundle(CommonBundle.javaClass, BUNDLE)
-
-    @JvmStatic
-    fun message(@PropertyKey(resourceBundle = BUNDLE) key: String, vararg params: Any) = bundle.getMessage(key, *params)
-
-    @Suppress("unused")
-    @JvmStatic
-    fun messagePointer(@PropertyKey(resourceBundle = BUNDLE) key: String, vararg params: Any) =
-        bundle.getLazyMessage(key, *params)
-}
