--- src/main/kotlin/works/szabope/plugins/common/action/ScrollToSourceDummyAction.kt
@@ -1,14 +1,0 @@
-package works.szabope.plugins.common.action
-
-import com.intellij.openapi.actionSystem.AnAction
-import com.intellij.openapi.actionSystem.AnActionEvent
-
-/**
- * This is a placeholder.
- * See init of [AbstractToolWindowPanel.addAutoScrollToSource][works.szabope.plugins.common.toolWindow.AbstractToolWindowPanel.addAutoScrollToSource]
- */
-class ScrollToSourceDummyAction : AnAction() {
-    override fun actionPerformed(e: AnActionEvent) {
-        // intentionally left blank
-    }
-}
