--- src/main/kotlin/works/szabope/plugins/common/messages/MessageConverter.kt
@@ -1,5 +1,0 @@
-package works.szabope.plugins.common.messages
-
-interface MessageConverter<in S, out T> {
-    fun convert(message: S): T
-}
