--- src/main/kotlin/works/szabope/plugins/common/services/BasicSettingsData.kt
@@ -1,8 +1,0 @@
-package works.szabope.plugins.common.services
-
-interface BasicSettingsData {
-    val executablePath: String?
-    val configFilePath: String?
-    val arguments: String?
-    val scanBeforeCheckIn: Boolean
-}
