--- src/main/kotlin/works/szabope/plugins/common/services/SettingsData.kt
@@ -1,7 +1,0 @@
-package works.szabope.plugins.common.services
-
-interface SettingsData : BasicSettingsData {
-    val useProjectSdk: Boolean
-    val projectDirectory: String?
-    val excludeNonProjectFiles: Boolean
-}
