--- src/main/kotlin/works/szabope/plugins/common/services/SeverityConfig.kt
@@ -1,5 +1,0 @@
-package works.szabope.plugins.common.services
-
-import javax.swing.Icon
-
-data class SeverityConfig(val level: String, val text: String, val description: String, val icon: Icon)
