--- src/main/kotlin/works/szabope/plugins/common/services/ToolResultItem.kt
@@ -1,3 +1,0 @@
-package works.szabope.plugins.common.services
-
-interface ToolResultItem
