--- src/main/kotlin/works/szabope/plugins/common/services/tool/ToolOutputHandler.kt
@@ -1,7 +1,0 @@
-package works.szabope.plugins.common.services.tool
-
-import kotlinx.coroutines.flow.Flow
-
-interface ToolOutputHandler {
-    suspend fun handle(stdout: Flow<String>): Result<Unit>
-}
